requirements.txt
python-telegram-bot
python-dotenv
git+https://github.com/an-dutra/pytube.git@patch-1
